Move to have base64url-encoding only done for serialization.
This corrects a few cases where base64url encoding
was done within an algorithm. For example, the
combined MAC representation uses the base64url-
encoded versions of the issuer protected header
and of the payloads, rather than their octet
string representations.

This is a compatibility break with the prior draft.

This also clarifies the appendix text to represent
examples are showing base64url-encoded data,
rather than the actual algorithmic data being
base64url-encoded.